schemadiff: move Difference next to Diff in diff.go
Difference is the return type of Diff, so co-locate them for discoverability.
difftypes.go now holds only the DiffType vocabulary (which grows as object-type
coverage expands) and no longer needs the schemasnapshot import. Same-package
move; no API or behavior change.
